+++ PathFind/Pathfinding.Domain.Interface/Repositories/IVerticesRepository.cs
@@ -19,7 +19,6 @@ namespace Pathfinding.Domain.Interface.Repositories
         Task<bool> UpdateVerticesAsync(IEnumerable<Vertex> vertices,
             CancellationToken token = default);
 
-        Task<Vertex> ReadAsync(int vertexId,
-            CancellationToken token = default);
+        Task<Vertex> ReadAsync(long vertexId, CancellationToken token = default);
     }
 }
